requirements.txt
flask
flask-login
flask-sqlalchemy
oauthlib
requests
python-dotenv
youtube_transcript_api
openai
deepgram-sdk>=2.11.0
google-generativeai>=0.3.0
websockets>=12.0
asyncio>=3.4.3
pyaudio>=0.2.13
python-dotenv>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
aiohttp>=3.9.0 